pyproject.toml
[tool.poetry]
name = "LlamaIndexAPI"
version = "0.0.1"
description = "A FastAPI starter point for Llama Index projects"
authors = ["Zachary Handley"]
# New attributes
license = "MIT"
readme = "README.md"
homepage = "https://zachhandley.com"
repository = "https://github.com/zachhandley/LlamaIndexAPI"
keywords = ["ai", "docker", "python", "python3.11", "llama-index", "fastapi"]
include = ["LICENSE"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
# Updated Python version
python = ">=3.11.4,<3.12"
# langchain = ">=0.0.250"
# pydantic = ">=2.0.0"
llama-index = "^0.9.26"
anthropic = "^0.3.10"
appwrite = "^2.0.2"
asyncio = "^3.4.3"
requests = "^2.31.0"
pypdf = "^3.15.2"
lxml = "^4.9.3"
sentence-transformers = "^2.2.2"
guidance = "^0.0.64"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
gunicorn = "^21.2.0"
trafilatura = "^1.6.1"
torch = ">=2.0.0, !=2.0.1"
metaphor-python = "^0.1.16"
nest-asyncio = "^1.5.7"
fastapi = "^0.99"
langchain = "^0.0.340"
ftfy = "^6.1.3"
regex = "^2023.12.25"
tqdm = "^4.66.1"
clip = {git = "https://github.com/openai/CLIP.git"}
matplotlib = "^3.8.2"
scikit-image = "^0.22.0"
torchvision = "^0.16.2"
qdrant-client = "^1.7.0"
pillow = "^10.2.0"
aiohttp = "^3.9.1"


[tool.poetry.dev-dependencies]
typing-inspect = "0.8.0"
typing_extensions = "4.5.0"
types-requests = "2.28.11.8"
isort = "5.11.4"
pytest = "7.3.1"
pytest-cov = "4.0.0"
coverage = "7.2.1"
black = "23.3.0"
pre-commit = "3.1.1"
semver = "2.13.0"
pylint = "2.16.2"
testcontainers = "3.7.1"
bcrypt = "4.0.1"

# https://docs.pytest.org/en/latest/reference/customize.html
[tool.pytest.ini_options]
addopts = "--ignore=tests/functionality/test_worker_reload.py"
testpaths = ["tests"]

# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
[tool.black]
line-length = 80
target-version = ['py311']

[tool.pylint.main]
errors-only = true
recursive = "y"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
